requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
pymysql